+++ packages/0x.js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v0.30.0 - _January 17, 2018_
+
+    * Add an error parameter to the order watcher callback (#312)
+    * Fix a bug making it impossible to catch some errors from awaitTransactionMinedAsync (#312)
+    * Fix a bug in fillOrdersUpTo validation making it impossible to fill up to if user doesn't have enough balance to fully fill all the orders (#321)
+
 ## v0.29.1 - _January 11, 2018_
 
     * Fixed bignumber config issue #301 (#305)