+++ packages/utils/CHANGELOG.md
@@ -1,3 +1,6 @@
 # CHANGELOG
 
-## vx.x.x
+## v0.2.0 - _January 17, 2018_
+
+* Add `onError` parameter to `intervalUtils.setAsyncExcludingInterval` (#312)
+* Add `intervalUtils.setInterval` (#312)